+++ resources/views/emails/password-reset.blade.php
@@ -4,30 +4,30 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 --}}
 
 <x-mail::message>
-# Reset Your Password
+# {{ __('Reset Your Password') }}
 
-Hello {{ $user->name }},
+{{ __('Hello') }} {{ $user->name }},
 
-You requested a password reset for your SecPal account.
+{{ __('You requested a password reset for your SecPal account.') }}
 
-Click the button below to reset your password. This link will expire in **{{ $expiresInMinutes }} minutes**.
+{{ __('Click the button below to reset your password. This link will expire in :minutes minutes.', ['minutes' => $expiresInMinutes]) }}
 
 <x-mail::button :url="$resetUrl">
-Reset Password
+{{ __('Reset Password') }}
 </x-mail::button>
 
-**Security Notice:**
-- Never share this link with anyone
-- If you didn't request this reset, please ignore this email
-- Your password will remain unchanged until you complete the reset process
+**{{ __('Security Notice:') }}**
+- {{ __('Never share this link with anyone') }}
+- {{ __('If you didn\'t request this reset, please ignore this email') }}
+- {{ __('Your password will remain unchanged until you complete the reset process') }}
 
-Thanks,<br>
+{{ __('Thanks') }},<br>
 {{ config('app.name') }}
 
 ---
 
 <small>
-If you're having trouble clicking the "Reset Password" button, copy and paste the URL below into your web browser:<br>
+{{ __('If you\'re having trouble clicking the ":button" button, copy and paste the URL below into your web browser:', ['button' => __('Reset Password')]) }}<br>
 {{ $resetUrl }}
 </small>
 </x-mail::message>